refactor($compile): 2단계 피드백 반영
2단계 피드백 반영
- LottoTicket getMatchedLottoNumberCount stream, filter, count를 활용하여 개선
- ResultView enum의 values 활용하여 개선
- LottoRank 값 캐시처리
- LottoNumber 사용하지 않는 생성자 제거
- LottoNumber 상수 접근제어자 default로 변경

+++ src/main/java/lotto/domain/LottoResult.java
@@ -4,9 +4,9 @@ import java.util.Objects;
 
 public class LottoResult {
 
-  private final int matchedCount;
+  private final long matchedCount;
 
-  public LottoResult(int matchedCount) {
+  public LottoResult(long matchedCount) {
     this.matchedCount = matchedCount;
   }
